chore: rename add_users to users
